+++ Project1/Objects/Actor.cs
@@ -146,7 +146,7 @@ namespace Start_a_Town_
             var old = need.Value;
             //need.Value = modOldValue(need.Value);
             need.SetValue(modOldValue(need.Value), this);
-            this.Net?.EventOccured(Message.Types.NeedUpdated, this, need, need.Value - old);
+            this.Net?.EventOccured((int)Message.Types.NeedUpdated, this, need, need.Value - old);
         }
 
         internal void Carry(Entity item)
